Performance & Bug Fixes
Improved the performance from ~ 370MBps to ~660MBps (MegaBytes a Second,
Not MegaBit)
Fixed quite a few bugs
Created OptimizedPayloadStream which will handle most traffic and make
it all nice
Fixed the TestServer Private/Public key locations
My Surface Pro2 can now  run SSP3

+++ SecureSocketProtocol3/Network/Messages/TCP/MsgCreateConnectionResponse.cs
@@ -11,12 +11,12 @@ namespace SecureSocketProtocol3.Network.Messages.TCP
     public class MsgCreateConnectionResponse : IMessage
     {
         [ProtoMember(1)]
-        public int ConnectionId;
+        public ushort ConnectionId;
 
         [ProtoMember(2)]
         public bool Success;
 
-        public MsgCreateConnectionResponse(int ConnectionId, bool Success)
+        public MsgCreateConnectionResponse(ushort ConnectionId, bool Success)
             : base()
         {
             this.ConnectionId = ConnectionId;
